chore: document modular refactor plan and extract text utilities
